+++ pseudosci/units/tests/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+# -*- coding:utf-8 -*-
+"""Tests unitaires des unités de mesure du projet."""
